SoftEng22-96-main/requirements.txt
pandas
mysql.connector
argparse
requests
